requirements.txt
rdflib
requests
python-dotenv